convert two slices into one
